organize code; add icmp test

+++ tvt_client.py
@@ -9,7 +9,7 @@ parser = argparse.ArgumentParser()
 parser.add_argument("host", type=str, help="host to connect")
 parser.add_argument("-p", "--port", type=int, help="set port number (default 12345)",
                     default=12345)
-parser.add_argument("-i", "--interval", type=int, help="set interval (default 1)",
+parser.add_argument("-i", "--interval", type=float, help="set interval (default 1)",
                     default=1)
 args = parser.parse_args()
 
@@ -25,3 +25,4 @@ try:
         time.sleep(args.interval)
 finally:
     sock.close()
+
